+++ Cargo.toml
@@ -4,12 +4,10 @@ version = "0.1.0"
 authors = ["rink1969 <rink1969@gmail.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-tonic = "0.2"
-prost = "0.6"
-tokio = { version = "0.2", features = ["macros"] }
+tonic = "*"
+prost = "*"
+tokio = { version = "*", features = ["full"] }
 
 [build-dependencies]
-tonic-build = "0.2.0"
+tonic-build = "*"
