--- dev_build.rs
@@ -1,19 +1,0 @@
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tonic_build::configure()
-        .build_client(true)
-        .build_server(true)
-        .format(true)
-        .out_dir("src")
-        .compile(
-            &["protos/controller.proto",
-                "protos/blockchain.proto",
-                "protos/common.proto",
-                "protos/consensus.proto",
-                "protos/executor.proto",
-                "protos/network.proto",
-                "protos/storage.proto",
-                "protos/kms.proto"],
-            &["protos"],
-        )?;
-    Ok(())
-}
