--- src/consensus.rs
@@ -1,231 +1,0 @@
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ConsensusConfiguration {
-    #[prost(uint32, tag = "1")]
-    pub block_interval: u32,
-    #[prost(bytes, repeated, tag = "2")]
-    pub validators: ::std::vec::Vec<std::vec::Vec<u8>>,
-}
-#[doc = r" Generated client implementations."]
-pub mod consensus_service_client {
-    #![allow(unused_variables, dead_code, missing_docs)]
-    use tonic::codegen::*;
-    pub struct ConsensusServiceClient<T> {
-        inner: tonic::client::Grpc<T>,
-    }
-    impl ConsensusServiceClient<tonic::transport::Channel> {
-        #[doc = r" Attempt to create a new client by connecting to a given endpoint."]
-        pub async fn connect<D>(dst: D) -> Result<Self, tonic::transport::Error>
-        where
-            D: std::convert::TryInto<tonic::transport::Endpoint>,
-            D::Error: Into<StdError>,
-        {
-            let conn = tonic::transport::Endpoint::new(dst)?.connect().await?;
-            Ok(Self::new(conn))
-        }
-    }
-    impl<T> ConsensusServiceClient<T>
-    where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
-        T::ResponseBody: Body + HttpBody + Send + 'static,
-        T::Error: Into<StdError>,
-        <T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,
-    {
-        pub fn new(inner: T) -> Self {
-            let inner = tonic::client::Grpc::new(inner);
-            Self { inner }
-        }
-        pub fn with_interceptor(inner: T, interceptor: impl Into<tonic::Interceptor>) -> Self {
-            let inner = tonic::client::Grpc::with_interceptor(inner, interceptor);
-            Self { inner }
-        }
-        pub async fn reconfigure(
-            &mut self,
-            request: impl tonic::IntoRequest<super::ConsensusConfiguration>,
-        ) -> Result<tonic::Response<super::super::common::SimpleResponse>, tonic::Status> {
-            self.inner.ready().await.map_err(|e| {
-                tonic::Status::new(
-                    tonic::Code::Unknown,
-                    format!("Service was not ready: {}", e.into()),
-                )
-            })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path =
-                http::uri::PathAndQuery::from_static("/consensus.ConsensusService/Reconfigure");
-            self.inner.unary(request.into_request(), path, codec).await
-        }
-        pub async fn check_block(
-            &mut self,
-            request: impl tonic::IntoRequest<super::super::common::ProposalWithProof>,
-        ) -> Result<tonic::Response<super::super::common::SimpleResponse>, tonic::Status> {
-            self.inner.ready().await.map_err(|e| {
-                tonic::Status::new(
-                    tonic::Code::Unknown,
-                    format!("Service was not ready: {}", e.into()),
-                )
-            })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path =
-                http::uri::PathAndQuery::from_static("/consensus.ConsensusService/CheckBlock");
-            self.inner.unary(request.into_request(), path, codec).await
-        }
-    }
-    impl<T: Clone> Clone for ConsensusServiceClient<T> {
-        fn clone(&self) -> Self {
-            Self {
-                inner: self.inner.clone(),
-            }
-        }
-    }
-    impl<T> std::fmt::Debug for ConsensusServiceClient<T> {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "ConsensusServiceClient {{ ... }}")
-        }
-    }
-}
-#[doc = r" Generated server implementations."]
-pub mod consensus_service_server {
-    #![allow(unused_variables, dead_code, missing_docs)]
-    use tonic::codegen::*;
-    #[doc = "Generated trait containing gRPC methods that should be implemented for use with ConsensusServiceServer."]
-    #[async_trait]
-    pub trait ConsensusService: Send + Sync + 'static {
-        async fn reconfigure(
-            &self,
-            request: tonic::Request<super::ConsensusConfiguration>,
-        ) -> Result<tonic::Response<super::super::common::SimpleResponse>, tonic::Status>;
-        async fn check_block(
-            &self,
-            request: tonic::Request<super::super::common::ProposalWithProof>,
-        ) -> Result<tonic::Response<super::super::common::SimpleResponse>, tonic::Status>;
-    }
-    #[derive(Debug)]
-    #[doc(hidden)]
-    pub struct ConsensusServiceServer<T: ConsensusService> {
-        inner: _Inner<T>,
-    }
-    struct _Inner<T>(Arc<T>, Option<tonic::Interceptor>);
-    impl<T: ConsensusService> ConsensusServiceServer<T> {
-        pub fn new(inner: T) -> Self {
-            let inner = Arc::new(inner);
-            let inner = _Inner(inner, None);
-            Self { inner }
-        }
-        pub fn with_interceptor(inner: T, interceptor: impl Into<tonic::Interceptor>) -> Self {
-            let inner = Arc::new(inner);
-            let inner = _Inner(inner, Some(interceptor.into()));
-            Self { inner }
-        }
-    }
-    impl<T, B> Service<http::Request<B>> for ConsensusServiceServer<T>
-    where
-        T: ConsensusService,
-        B: HttpBody + Send + Sync + 'static,
-        B::Error: Into<StdError> + Send + 'static,
-    {
-        type Response = http::Response<tonic::body::BoxBody>;
-        type Error = Never;
-        type Future = BoxFuture<Self::Response, Self::Error>;
-        fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-            Poll::Ready(Ok(()))
-        }
-        fn call(&mut self, req: http::Request<B>) -> Self::Future {
-            let inner = self.inner.clone();
-            match req.uri().path() {
-                "/consensus.ConsensusService/Reconfigure" => {
-                    #[allow(non_camel_case_types)]
-                    struct ReconfigureSvc<T: ConsensusService>(pub Arc<T>);
-                    impl<T: ConsensusService>
-                        tonic::server::UnaryService<super::ConsensusConfiguration>
-                        for ReconfigureSvc<T>
-                    {
-                        type Response = super::super::common::SimpleResponse;
-                        type Future = BoxFuture<tonic::Response<Self::Response>, tonic::Status>;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::ConsensusConfiguration>,
-                        ) -> Self::Future {
-                            let inner = self.0.clone();
-                            let fut = async move { inner.reconfigure(request).await };
-                            Box::pin(fut)
-                        }
-                    }
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let interceptor = inner.1.clone();
-                        let inner = inner.0;
-                        let method = ReconfigureSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = if let Some(interceptor) = interceptor {
-                            tonic::server::Grpc::with_interceptor(codec, interceptor)
-                        } else {
-                            tonic::server::Grpc::new(codec)
-                        };
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/consensus.ConsensusService/CheckBlock" => {
-                    #[allow(non_camel_case_types)]
-                    struct CheckBlockSvc<T: ConsensusService>(pub Arc<T>);
-                    impl<T: ConsensusService>
-                        tonic::server::UnaryService<super::super::common::ProposalWithProof>
-                        for CheckBlockSvc<T>
-                    {
-                        type Response = super::super::common::SimpleResponse;
-                        type Future = BoxFuture<tonic::Response<Self::Response>, tonic::Status>;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::super::common::ProposalWithProof>,
-                        ) -> Self::Future {
-                            let inner = self.0.clone();
-                            let fut = async move { inner.check_block(request).await };
-                            Box::pin(fut)
-                        }
-                    }
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let interceptor = inner.1.clone();
-                        let inner = inner.0;
-                        let method = CheckBlockSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = if let Some(interceptor) = interceptor {
-                            tonic::server::Grpc::with_interceptor(codec, interceptor)
-                        } else {
-                            tonic::server::Grpc::new(codec)
-                        };
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                _ => Box::pin(async move {
-                    Ok(http::Response::builder()
-                        .status(200)
-                        .header("grpc-status", "12")
-                        .body(tonic::body::BoxBody::empty())
-                        .unwrap())
-                }),
-            }
-        }
-    }
-    impl<T: ConsensusService> Clone for ConsensusServiceServer<T> {
-        fn clone(&self) -> Self {
-            let inner = self.inner.clone();
-            Self { inner }
-        }
-    }
-    impl<T: ConsensusService> Clone for _Inner<T> {
-        fn clone(&self) -> Self {
-            Self(self.0.clone(), self.1.clone())
-        }
-    }
-    impl<T: std::fmt::Debug> std::fmt::Debug for _Inner<T> {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{:?}", self.0)
-        }
-    }
-    impl<T: ConsensusService> tonic::transport::NamedService for ConsensusServiceServer<T> {
-        const NAME: &'static str = "consensus.ConsensusService";
-    }
-}
